+++ markets/spot-market/contracts/storage/AsyncOrderClaim.sol
@@ -108,8 +108,8 @@ library AsyncOrderClaim {
         Data storage claim,
         SettlementStrategy.Data storage settlementStrategy
     ) internal view {
-        uint startTime = claim.settlementTime;
-        uint expirationTime = startTime + settlementStrategy.settlementWindowDuration;
+        uint256 startTime = claim.settlementTime;
+        uint256 expirationTime = startTime + settlementStrategy.settlementWindowDuration;
 
         if (block.timestamp < startTime || block.timestamp >= expirationTime) {
             revert OutsideSettlementWindow(block.timestamp, startTime, expirationTime);
@@ -120,7 +120,7 @@ library AsyncOrderClaim {
         Data storage claim,
         SettlementStrategy.Data storage settlementStrategy
     ) internal view {
-        uint expirationTime = claim.settlementTime + settlementStrategy.settlementWindowDuration;
+        uint256 expirationTime = claim.settlementTime + settlementStrategy.settlementWindowDuration;
 
         if (block.timestamp < expirationTime) {
             revert IneligibleForCancellation(block.timestamp, expirationTime);
